mixed_cg_merge: do not use SSE in single precision operators because there is no SSE implementation
