+++ operator/D_psi_32.c
@@ -31,6 +31,22 @@
 # include<config.h>
 #endif
 
+// work-around for missing single precision implementation of inline SSE
+#ifdef SSE
+#define REDEFSSE
+#undef SSE
+#endif
+
+#ifdef SSE2
+#define REDEFSSE2
+#undef SSE2
+#endif
+
+#ifdef SSE3
+#define REDEFSSE3
+#undef SSE3
+#endif
+
 #include <stdlib.h>
 #include <stdio.h>
 #include <math.h>
@@ -375,3 +391,17 @@ _Complex float ALIGN32 phase_3_32 = (_Complex float) phase_3;
 #endif
 }
 
+#ifdef REDEFSSE
+#undef REDEFSSE
+#define SSE
+#endif
+
+#ifdef REDEFSSE2
+#undef REDEFSSE2
+#define SSE2
+#endif
+
+#ifdef REDEFSSE3
+#undef REDEFSSE3
+#define SSE3
+#endif
