+++ operator/Hopping_Matrix_32.c
@@ -50,6 +50,23 @@
 #ifdef HAVE_CONFIG_H
 # include<config.h>
 #endif
+
+// work-around for missing single precision implementation of inline SSE
+#ifdef SSE
+#define REDEFSSE
+#undef SSE
+#endif
+
+#ifdef SSE2
+#define REDEFSSE2
+#undef SSE2
+#endif
+
+#ifdef SSE3
+#define REDEFSSE3
+#undef SSE3
+#endif
+
 #include <stdlib.h>
 #include <stdio.h>
 #ifdef OMP
@@ -110,4 +127,19 @@ void Hopping_Matrix_32(const int ieo, spinor32 * const l, spinor32 * const k) {
 #endif
   return;
 }
+
+#ifdef REDEFSSE
+#undef REDEFSSE
+#define SSE
+#endif
+
+#ifdef REDEFSSE2
+#undef REDEFSSE2
+#define SSE2
+#endif
+
+#ifdef REDEFSSE3
+#undef REDEFSSE3
+#define SSE3
+#endif
