+++ operator/Makefile.in
@@ -31,7 +31,8 @@ COMPILE = ${CC} $(DEFS) ${INCLUDES} ${CFLAGS}
 
 LIBRARIES = liboperator
 liboperator_TARGETS = clover_accumulate_deriv clover_deriv clovertm_operators clover_leaf \
-  tm_operators_nd tm_operators_nd_32 clover_term clover_invert clover_det
+  tm_operators_nd tm_operators_nd_32 clover_term clover_invert clover_det \
+	clovertm_operators_32
 
 liboperator_STARGETS = Hopping_Matrix_nocom tm_times_Hopping_Matrix Hopping_Matrix Hopping_Matrix_32 Hopping_Matrix_32_nocom \
 	tm_operators tm_operators_32 tm_sub_Hopping_Matrix D_psi D_psi_32 Dov_psi Dov_proj
